requirements.txt
streamlit
opencv-python-headless
numpy<2.0.0
tensorflow-cpu==2.15.0
Pillow